chore: resolve CodeFactor findings — builtin shadowing and cross-file duplication
Rename identifiers shadowing Go builtins (real in helm_values and three
test sites, copy in appset discovery and loader tests). Consolidate
duplicated code: shared pluginexec process types/helpers into an
untagged file, git URL classification and error redaction into a new
internal/giturl package, kustomize build token normalization into
render.NormalizeKustomizeBuildTokens, the home-ops smoke scripts'
cleanup/detect/update helpers into scripts/lib/home-ops-smoke.sh (update
strictness parameterized, behavior verified fixture-by-fixture against
the originals), and the duplicate .header-actions media-query rule in
the diff HTML template plus its generated example.
