+++ internal/pluginexec/process_unix.go
@@ -5,30 +5,9 @@ package pluginexec
 import (
 	"context"
 	"errors"
-	"io"
 	"os/exec"
 	"syscall"
-	"time"
 )
-
-type processRequest struct {
-	Path    string
-	Args    []string
-	Dir     string
-	Env     []string
-	Timeout time.Duration
-	Stdin   io.Reader
-	Stdout  io.Writer
-	Stderr  io.Writer
-}
-
-type exitError struct {
-	Code int
-}
-
-func (e exitError) Error() string {
-	return "process exited"
-}
 
 func runProcess(ctx context.Context, request processRequest) error {
 	runCtx := ctx
@@ -70,15 +49,4 @@ func runProcess(ctx context.Context, request processRequest) error {
 		}
 		return runCtx.Err()
 	}
-}
-
-func processWaitError(err error) error {
-	if err == nil {
-		return nil
-	}
-	var exit *exec.ExitError
-	if errors.As(err, &exit) {
-		return exitError{Code: exit.ExitCode()}
-	}
-	return err
 }
