+++ internal/pluginexec/process_windows.go
@@ -5,29 +5,8 @@ package pluginexec
 import (
 	"context"
 	"errors"
-	"io"
 	"os/exec"
-	"time"
 )
-
-type processRequest struct {
-	Path    string
-	Args    []string
-	Dir     string
-	Env     []string
-	Timeout time.Duration
-	Stdin   io.Reader
-	Stdout  io.Writer
-	Stderr  io.Writer
-}
-
-type exitError struct {
-	Code int
-}
-
-func (e exitError) Error() string {
-	return "process exited"
-}
 
 func runProcess(ctx context.Context, request processRequest) error {
 	runCtx := ctx
@@ -67,15 +46,4 @@ func runProcess(ctx context.Context, request processRequest) error {
 		}
 		return runCtx.Err()
 	}
-}
-
-func processWaitError(err error) error {
-	if err == nil {
-		return nil
-	}
-	var exit *exec.ExitError
-	if errors.As(err, &exit) {
-		return exitError{Code: exit.ExitCode()}
-	}
-	return err
 }
