+++ internal/render/kustomize_build_options.go
@@ -2,6 +2,7 @@ package render
 
 import (
 	"fmt"
+	"path/filepath"
 	"strings"
 
 	"sigs.k8s.io/kustomize/api/types"
@@ -26,6 +27,49 @@ func defaultKustomizeBuildSettings() kustomizeBuildSettings {
 func ValidateKustomizeBuildOptions(options []string) error {
 	_, err := parseKustomizeBuildOptions(options)
 	return err
+}
+
+// NormalizeKustomizeBuildTokens validates that tokens form a plain
+// "kustomize build" invocation and returns just the supported build options,
+// rejecting unsupported flags, paths, and remote operands.
+func NormalizeKustomizeBuildTokens(tokens []string) ([]string, error) {
+	if len(tokens) < 2 || filepath.Base(tokens[0]) != "kustomize" || tokens[1] != "build" {
+		return nil, fmt.Errorf("generate command is not kustomize build")
+	}
+	options := make([]string, 0, len(tokens)-2)
+	for i := 2; i < len(tokens); i++ {
+		token := strings.TrimSpace(tokens[i])
+		if token == "" || token == "." {
+			continue
+		}
+		switch {
+		case token == "--enable-helm",
+			token == "--enable-alpha-plugins",
+			token == "--enable-exec",
+			strings.HasPrefix(token, "--helm-api-versions="),
+			strings.HasPrefix(token, "--load-restrictor="):
+			options = append(options, token)
+		case token == "--helm-api-versions" || token == "--load-restrictor":
+			if i+1 >= len(tokens) {
+				return nil, fmt.Errorf("kustomize build option requires a value")
+			}
+			i++
+			value := strings.TrimSpace(tokens[i])
+			if value == "" || strings.HasPrefix(value, "-") {
+				return nil, fmt.Errorf("kustomize build option requires a value")
+			}
+			options = append(options, token, value)
+		default:
+			if strings.HasPrefix(token, "-") {
+				return nil, fmt.Errorf("unsupported kustomize build option")
+			}
+			return nil, fmt.Errorf("unsupported kustomize build path or remote operand")
+		}
+	}
+	if err := ValidateKustomizeBuildOptions(options); err != nil {
+		return nil, fmt.Errorf("unsupported kustomize build option")
+	}
+	return options, nil
 }
 
 func parseKustomizeBuildOptions(options []string) (kustomizeBuildSettings, error) {
