+++ scripts/home-ops-pattern-smoke.sh
@@ -11,34 +11,8 @@ BASELINE=""
 CURRENT=""
 REMOTE_CACHE=""
 
-cleanup() {
-  local status=$?
-  local cleanup_status=0
-  if [[ -n "${BASELINE}" && -d "${BASELINE}" ]]; then
-    if ! git -C "${ROOT}" worktree remove --force "${BASELINE}" >&2; then
-      echo "failed to remove baseline worktree: ${BASELINE}" >&2
-      cleanup_status=1
-    fi
-  fi
-  if [[ -n "${CURRENT}" && -d "${CURRENT}" ]]; then
-    if ! git -C "${ROOT}" worktree remove --force "${CURRENT}" >&2; then
-      echo "failed to remove current worktree: ${CURRENT}" >&2
-      cleanup_status=1
-    fi
-  fi
-  if [[ "${cleanup_status}" -eq 0 && -n "${TMP_DIR}" && -d "${TMP_DIR}" ]]; then
-    rm -rf "${TMP_DIR}"
-  fi
-  if [[ "${cleanup_status}" -ne 0 ]]; then
-    if [[ -n "${TMP_DIR}" ]]; then
-      echo "leaving temporary smoke directory for inspection: ${TMP_DIR}" >&2
-    fi
-    if [[ "${status}" -eq 0 ]]; then
-      exit 2
-    fi
-  fi
-  exit "${status}"
-}
+# shellcheck source=scripts/lib/home-ops-smoke.sh
+source "${SCRIPT_DIR}/lib/home-ops-smoke.sh"
 trap cleanup EXIT
 
 fail() {
@@ -84,91 +58,6 @@ replace_once_literal() {
   fi
 }
 
-detect_helm_chart_version() {
-  local file="$1"
-  local chart_name="$2"
-
-  awk -v chart_name="${chart_name}" '
-    function indent(line) {
-      match(line, /^[[:space:]]*/)
-      return RLENGTH
-    }
-    function clean_value(line) {
-      sub(/[[:space:]]*#.*/, "", line)
-      sub(/^[[:space:]]*/, "", line)
-      sub(/[[:space:]]*$/, "", line)
-      gsub(/^"|"$/, "", line)
-      return line
-    }
-    function reset_item() {
-      in_item = 0
-      item_name = ""
-      item_version = ""
-    }
-    function read_inline_field(line, field) {
-      sub("^[[:space:]]*-[[:space:]]*" field ":[[:space:]]*", "", line)
-      return clean_value(line)
-    }
-    function read_child_field(line, field) {
-      sub("^[[:space:]]*" field ":[[:space:]]*", "", line)
-      return clean_value(line)
-    }
-    function finish_item() {
-      if (!found && in_item && item_name == chart_name && item_version != "") {
-        print item_version
-        found = 1
-      }
-    }
-    indent($0) == 0 && /^[[:space:]]*helmCharts:[[:space:]]*($|#)/ {
-      in_helm = 1
-      helm_indent = indent($0)
-      chart_item_indent = -1
-      reset_item()
-      next
-    }
-    in_helm && $0 !~ /^[[:space:]]*($|#)/ && indent($0) <= helm_indent {
-      finish_item()
-      in_helm = 0
-      reset_item()
-    }
-    !in_helm {
-      next
-    }
-    chart_item_indent < 0 && indent($0) > helm_indent && /^[[:space:]]*-/ {
-      chart_item_indent = indent($0)
-    }
-    chart_item_indent >= 0 && indent($0) == chart_item_indent && /^[[:space:]]*-/ {
-      finish_item()
-      if (found) {
-        exit
-      }
-      reset_item()
-      in_item = 1
-      if ($0 ~ /^[[:space:]]*-[[:space:]]*name:[[:space:]]*/) {
-        item_name = read_inline_field($0, "name")
-      }
-      if ($0 ~ /^[[:space:]]*-[[:space:]]*version:[[:space:]]*/) {
-        item_version = read_inline_field($0, "version")
-      }
-      next
-    }
-    in_item && indent($0) == chart_item_indent + 2 && /^[[:space:]]*name:[[:space:]]*/ {
-      item_name = read_child_field($0, "name")
-      next
-    }
-    in_item && indent($0) == chart_item_indent + 2 && /^[[:space:]]*version:[[:space:]]*/ {
-      item_version = read_child_field($0, "version")
-      next
-    }
-    END {
-      finish_item()
-      if (!found) {
-        exit 1
-      }
-    }
-  ' "${file}"
-}
-
 previous_semver() {
   local version="$1"
   local major
@@ -196,108 +85,7 @@ update_helm_chart_version_once() {
   local output="${file}.tmp"
 
   require_file "${file}"
-  awk -v chart_name="${chart_name}" -v version_to="${version_to}" '
-    function indent(line) {
-      match(line, /^[[:space:]]*/)
-      return RLENGTH
-    }
-    function clean_value(line) {
-      sub(/[[:space:]]*#.*/, "", line)
-      sub(/^[[:space:]]*/, "", line)
-      sub(/[[:space:]]*$/, "", line)
-      gsub(/^"|"$/, "", line)
-      return line
-    }
-    function reset_item() {
-      delete item_lines
-      item_count = 0
-      in_item = 0
-      item_name = ""
-      version_line = 0
-    }
-    function add_item_line(line) {
-      item_count++
-      item_lines[item_count] = line
-    }
-    function read_inline_field(line, field) {
-      sub("^[[:space:]]*-[[:space:]]*" field ":[[:space:]]*", "", line)
-      return clean_value(line)
-    }
-    function read_child_field(line, field) {
-      sub("^[[:space:]]*" field ":[[:space:]]*", "", line)
-      return clean_value(line)
-    }
-    function emit_item(   i, prefix) {
-      if (in_item && item_name == chart_name) {
-        match_count++
-        if (version_line > 0 && match_count == 1) {
-          prefix = item_lines[version_line]
-          sub(/version:.*/, "version: ", prefix)
-          item_lines[version_line] = prefix version_to
-          changed = 1
-        }
-      }
-      for (i = 1; i <= item_count; i++) {
-        print item_lines[i]
-      }
-      reset_item()
-    }
-    indent($0) == 0 && /^[[:space:]]*helmCharts:[[:space:]]*($|#)/ {
-      in_helm = 1
-      helm_indent = indent($0)
-      chart_item_indent = -1
-      reset_item()
-      print
-      next
-    }
-    in_helm && $0 !~ /^[[:space:]]*($|#)/ && indent($0) <= helm_indent {
-      emit_item()
-      in_helm = 0
-      print
-      next
-    }
-    in_helm && chart_item_indent < 0 && indent($0) > helm_indent && /^[[:space:]]*-/ {
-      chart_item_indent = indent($0)
-    }
-    in_helm && chart_item_indent >= 0 && indent($0) == chart_item_indent && /^[[:space:]]*-/ {
-      emit_item()
-      in_item = 1
-      add_item_line($0)
-      if ($0 ~ /^[[:space:]]*-[[:space:]]*name:[[:space:]]*/) {
-        item_name = read_inline_field($0, "name")
-      }
-      if ($0 ~ /^[[:space:]]*-[[:space:]]*version:[[:space:]]*/) {
-        version_line = item_count
-      }
-      next
-    }
-    in_helm && chart_item_indent >= 0 && indent($0) == chart_item_indent + 2 && /^[[:space:]]*name:[[:space:]]*/ {
-      add_item_line($0)
-      item_name = read_child_field($0, "name")
-      next
-    }
-    in_helm && chart_item_indent >= 0 && indent($0) == chart_item_indent + 2 && /^[[:space:]]*version:[[:space:]]*/ {
-      add_item_line($0)
-      version_line = item_count
-      next
-    }
-    in_helm && in_item {
-      add_item_line($0)
-      next
-    }
-    { print }
-    END {
-      emit_item()
-      if (match_count != 1) {
-        printf("expected exactly one top-level helmCharts entry named %s, found %d\n", chart_name, match_count) > "/dev/stderr"
-        exit 1
-      }
-      if (!changed) {
-        printf("chart %s did not have an editable top-level version field\n", chart_name) > "/dev/stderr"
-        exit 1
-      }
-    }
-  ' "${file}" > "${output}" || fail "failed to update chart ${chart_name} in ${file}"
+  update_helm_chart_version "${file}" "${chart_name}" "${version_to}" "${output}" 1 || fail "failed to update chart ${chart_name} in ${file}"
   mv "${output}" "${file}"
 }
