+++ site/static/examples/full-rendered-diff-view.html
@@ -707,12 +707,10 @@ body[data-view="unified"] .diff-table.side-by-side {
 	}
 	.header-actions {
 		grid-column: 3;
+		gap: 6px;
 	}
 	.report-header h1 {
 		font-size: 16px;
-	}
-	.header-actions {
-		gap: 6px;
 	}
 	.brand-logo {
 		width: 38px;
